# internal/httpapi/gen/types.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package gen

import (
	"time"
)

// CreateTaskRequest defines model for CreateTaskRequest.
type CreateTaskRequest struct {
	IsDone bool   `json:"is_done"`
	Task   string `json:"task"`
	UserId int64  `json:"user_id"`
}

// CreateTaskResponse defines model for CreateTaskResponse.
type CreateTaskResponse = Task

// CreateUserRequest defines model for CreateUserRequest.
type CreateUserRequest struct {
	Email    string `json:"email"`
	Password string `json:"password"`
}

// CreateUserResponse defines model for CreateUserResponse.
type CreateUserResponse = User

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Message string `json:"message"`
}

// ListTasksResponse defines model for ListTasksResponse.
type ListTasksResponse = []Task

// ListUsersResponse defines model for ListUsersResponse.
type ListUsersResponse = []User

// PatchTaskRequest defines model for PatchTaskRequest.
type PatchTaskRequest struct {
	IsDone *bool   `json:"is_done,omitempty"`
	Task   *string `json:"task,omitempty"`
}

// PatchTaskResponse defines model for PatchTaskResponse.
type PatchTaskResponse = Task

// PatchUserRequest defines model for PatchUserRequest.
type PatchUserRequest struct {
	Email    *string `json:"email,omitempty"`
	Password *string `json:"password,omitempty"`
}

// PatchUserResponse defines model for PatchUserResponse.
type PatchUserResponse = User

// Task defines model for Task.
type Task struct {
	Id     int64  `json:"id"`
	IsDone bool   `json:"is_done"`
	Task   string `json:"task"`
	UserId int64  `json:"user_id"`
}

// User defines model for User.
type User struct {
	CreatedAt time.Time `json:"created_at"`
	Email     string    `json:"email"`
	Id        int64     `json:"id"`
	Password  string    `json:"password"`
	UpdatedAt time.Time `json:"updated_at"`
}

// PostTaskJSONRequestBody defines body for PostTask for application/json ContentType.
type PostTaskJSONRequestBody = CreateTaskRequest

// PatchTaskJSONRequestBody defines body for PatchTask for application/json ContentType.
type PatchTaskJSONRequestBody = PatchTaskRequest

// PostUserJSONRequestBody defines body for PostUser for application/json ContentType.
type PostUserJSONRequestBody = CreateUserRequest

// PatchUserByIDJSONRequestBody defines body for PatchUserByID for application/json ContentType.
type PatchUserByIDJSONRequestBody = PatchUserRequest