# openapi.yaml
openapi: 3.0.3
info:
  title: Task API
  version: 1.0.0
paths:
  /tasks:
    get:
      tags:
        - tasks
      operationId: listTasks
      summary: List tasks
      responses:
        "200":
          description: Tasks list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListTasksResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags:
        - tasks
      operationId: PostTask
      summary: Create task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateTaskResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tasks/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          minimum: 1
    patch:
      tags:
        - tasks
      operationId: patchTask
      summary: Patch task by id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchTaskRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatchTaskResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - tasks
      operationId: deleteTask
      summary: Delete task by id
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users:
    get:
      tags:
        - users
      operationId: GetUsers
      summary: List users
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListUsersResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags:
        - users
      operationId: PostUser
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateUserResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}/tasks:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          minimum: 1
    get:
      tags:
        - tasks
      operationId: GetTasksByUserID
      summary: List tasks by user id
      responses:
        "200":
          description: Tasks list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListTasksResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          minimum: 1
    patch:
      tags:
        - users
      operationId: PatchUserByID
      summary: Patch user by id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchUserRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatchUserResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - users
      operationId: DeleteUserByID
      summary: Delete user by id
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    Task:
      type: object
      required:
        - id
        - task
        - is_done
        - user_id
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        task:
          type: string
        is_done:
          type: boolean
        user_id:
          type: integer
          format: int64
          minimum: 1
    CreateTaskRequest:
      type: object
      required:
        - task
        - is_done
        - user_id
      properties:
        task:
          type: string
        is_done:
          type: boolean
        user_id:
          type: integer
          format: int64
          minimum: 1
    PatchTaskRequest:
      type: object
      properties:
        task:
          type: string
        is_done:
          type: boolean
    CreateTaskResponse:
      $ref: "#/components/schemas/Task"
    PatchTaskResponse:
      $ref: "#/components/schemas/Task"
    ListTasksResponse:
      type: array
      items:
        $ref: "#/components/schemas/Task"
    User:
      type: object
      required:
        - id
        - email
        - password
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        email:
          type: string
        password:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateUserRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
        password:
          type: string
    PatchUserRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
    CreateUserResponse:
      $ref: "#/components/schemas/User"
    PatchUserResponse:
      $ref: "#/components/schemas/User"
    ListUsersResponse:
      type: array
      items:
        $ref: "#/components/schemas/User"
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string